Added more things to answer.

+++ Hello.py
@@ -3,8 +3,8 @@
 name = input("Hi, what's your name? ")
 print("Nice to meet you, " + name + "!")
 
-age = float(input("How old are you? ")
+age = int(input("How old are you? "))
 if age > 50:
-print("You're too old! ")
+  print("You're too old! ")
 else:
-print("You're too young! ")
+  print("You're too young! ")
